(PC-5221) move RootNavigator
